+++ packages/core/src/components/graphs/line.ts
@@ -42,11 +42,10 @@ export class Line extends Component {
 
 		// D3 line generator function
 		this.lineGenerator = line()
-			.x((d, i) => this.services.axes.getXValue(d, i))
-			.y((d, i) => this.services.axes.getYValue(d, i))
+			.x((d, i) => this.services.cartesianScales.getDomainValue(d, i))
+			.y((d, i) => this.services.cartesianScales.getRangeValue(d, i))
 			.curve(this.services.curves.getD3Curve())
-			// Describe which parts of the line domain are defined
-			.defined((d: any, i) => {
+      .defined((d: any, i) => {
 				if (!d || d.value === null) {
 					return false;
 				}