+++ packages/core/src/services/index.ts
@@ -4,5 +4,5 @@ export * from "./essentials/events";
 export * from "./essentials/transitions";
 
 // MISC
-export * from "./axes";
+export * from "./scales-cartesian";
 export * from "./curves";