+++ packages/core/src/tests/test-environment.ts
@@ -3,64 +3,14 @@ import { Chart } from "../chart";
 import * as Charts from "../index";
 import { createChartHolder } from "./tools";
 
-export const data = {
-	labels: ["Qty", "More", "Sold", "Restocking", "Misc"],
-	datasets: [
-		{
-			label: "Dataset 1",
-			fillColors: [Charts.defaultColors[0]],
-			data: [
-				2000,
-				4200,
-				7000,
-				4000,
-				19000
-			]
-		},
-		{
-			label: "Dataset 2",
-			fillColors: [Charts.defaultColors[1]],
-			data: [
-				0,
-				10000,
-				20000,
-				30000,
-				40000
-			]
-		},
-		{
-			label: "Dataset 3",
-			fillColors: [Charts.defaultColors[2]],
-			data: [
-				0,
-				20000,
-				40000,
-				60000,
-				80000
-			]
-		}
-	]
-};
+import { ChartData } from "../interfaces";
 
-export const options = {
-	axes: {
-		bottom: {
-			title: "2018 Annual Sales Figures",
-			type: "labels",
-			primary: true
-		},
-		left: {
-			secondary: true
-		},
-		top: {
-			title: "Dollars (CAD)"
-		}
-	},
-	legendClickable: true,
-	resizable: true,
-	height: "500px",
+import { groupedBarData, groupedBarOptions } from "../../demo/demo-data";
+
+export const data = groupedBarData as ChartData;
+export const options = Object.assign(groupedBarOptions, {
 	title: "My chart"
-};
+}) as any;
 
 export class TestEnvironment {
 	chartOptions = options;